requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
clinner==1.12.3
colorama==0.4.4
colorlog==3.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.7
djangorestframework==3.12.2
drf-yasg==1.20.0
gitdb==4.0.5
GitPython==3.1.14
health-check==3.4.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.9
py-healthcheck==1.10.1
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
smmap==3.0.5
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
